% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangle.R
\name{hypotenuse}
\alias{hypotenuse}
\title{Hypotenuse Calculator}
\usage{
hypotenuse(x)
}
\arguments{
\item{x}{an rtriangle type object}
}
\value{
numeric
}
\description{
This contrived function computes the hypotenuse of an 'rtriangle' class object
}

