% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangle.R
\name{perimeter}
\alias{perimeter}
\title{Perimeter Calculator}
\usage{
perimeter(x)
}
\arguments{
\item{x}{an rtriangle type object}
}
\value{
numeric
}
\description{
This contrived function computes the perimeter of an 'rtriangle' class object
}
\examples{
x <- rtriangle(3,4); perimeter(x)
}

