% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtriangle-package.r, R/triangle.R
\docType{package}
\name{rtriangle}
\alias{rtriangle}
\alias{rtriangle-package}
\title{rtriangle: If you ever needed to calculate the hypotenuse, perimeter, or area of a basic triangle}
\usage{
rtriangle(base, height)
}
\arguments{
\item{base}{a positive numeric}

\item{height}{a positive numeric}
}
\value{
rtriangle class object
}
\description{
Contrived function that computes the hypotenuse, perimeter, and area of a triangle.

This contrived function creates to 'rtriangle' class objects
}
\examples{
x <- rtriangle(3,4); class(x)
}

